refactor: split api/client.ts into domain-specific modules with backwards-compat re-exports
